requirements.txt
pytest==8.3.3
azure-core==1.31.0
azure-cosmos==4.7.0
azure-storage-blob==12.23.1
email_validator==2.2.0
fastapi==0.115.4
python-dotenv==1.0.1
starlette==0.41.2
uvicorn==0.32.0
